feat: migrate editable table to mui (#995)
* chore: update table

* chore: few tweaks

* fix: align BulkEditTable with existing actions/sort conventions and fix bulk-save data corruption

- Use actions.edit.onClick/actions.delete.onClick (passing the full row)
  instead of a separate handleDeleteRow prop and a hardcoded history.push,
  matching the convention used by every other list page in this codebase.
- Drop the SORT_ASCENDING/SORT_DESCENDING constants in favor of literal
  1/-1, matching openstack-uicore-foundation's own MUI table.
- Remove the unused afterUpdate mechanism (dead since media_uploads
  editing was removed pre-migration).
- Rework formatEventData to spread the raw event instead of overwriting
  real API fields with display strings under the same key. Since
  BulkEditTable now sends formatEventData's output straight to
  bulkUpdateEvents, the previous behavior corrupted start_date, end_date,
  duration, level, streaming_url/meeting_url/etherpad_link,
  streaming_type, and location on every bulk save, and could crash on
  tags. Display-only values now live under new keys (speaker_names,
  track_name, *_display) wired up via column render functions.
- Move formatEventData out of summitUtils.js into summit-event-list-page
  (its only consumer) and relocate its tests accordingly.


* chore: removing dead code/features

* chore: more fixes and refactors

* chore: move BulkEditTable to uicore

* fix: update BulkEditTable mock path after move to uicore


---------
