+++ src/actions/event-actions.js
@@ -51,7 +51,6 @@ import {
   FIVE_PER_PAGE,
   HOUR_AND_HALF
 } from "../utils/constants";
-import { getIdValue } from "../utils/summitUtils";
 
 URI.escapeQuerySpace = false;
 
@@ -195,12 +194,12 @@ export const normalizeBulkEvents = (entity) => {
     const normalizedEvent = {
       id: e.id,
       title: e.title,
-      selection_plan_id: getIdValue(e.selection_plan) || e.selection_plan_id,
+      selection_plan_id: e.selection_plan?.id || e.selection_plan_id,
       location_id: e.location?.id || e.location_id,
       start_date: e.start_date,
       end_date: e.end_date,
-      type_id: getIdValue(e.type) || e.type_id,
-      track_id: getIdValue(e.track) || e.track_id,
+      type_id: e.type?.id || e.type_id,
+      track_id: e.track?.id || e.track_id,
       duration: e.duration,
       streaming_url: e.streaming_url,
       streaming_type: e.streaming_type,
@@ -299,66 +298,65 @@ export const getEvents =
     });
   };
 
-export const bulkUpdateEvents =
-  (summitId, events) => async (dispatch, getState) => {
-    const { currentSummitState } = getState();
-    const accessToken = await getAccessTokenSafely();
-    const { currentSummit } = currentSummitState;
-    dispatch(startLoading());
+export const bulkUpdateEvents = (events) => async (dispatch, getState) => {
+  const { currentSummitState } = getState();
+  const accessToken = await getAccessTokenSafely();
+  const { currentSummit } = currentSummitState;
+  dispatch(startLoading());
 
-    const normalizedEvents = normalizeBulkEvents(
-      events.map((event) =>
-        normalizeEvent(
-          event,
-          currentSummit.event_types.find((et) => et.id === event.type_id)
-        )
+  const normalizedEvents = normalizeBulkEvents(
+    events.map((event) =>
+      normalizeEvent(
+        event,
+        currentSummit.event_types.find((et) => et.id === event.type_id)
       )
-    );
+    )
+  );
 
-    return putRequest(
-      null,
-      createAction(UPDATED_REMOTE_EVENTS)({}),
-      `${window.API_BASE_URL}/api/v1/summits/${summitId}/events/?access_token=${accessToken}`,
-      {
-        events: normalizedEvents
-      },
-      authErrorHandler
-    )({})(dispatch)
-      .then(() => {
-        dispatch(stopLoading());
-        dispatch(
-          showSuccessMessage(
-            T.translate("bulk_actions_page.messages.update_success"),
-            () => history.push(`/app/summits/${currentSummit.id}/events/`)
-          )
-        );
-        const {
-          currentEventListState: {
-            term,
-            currentPage,
-            perPage,
-            order,
-            orderDir,
-            filters,
-            extraColumns
-          }
-        } = getState();
-        dispatch(
-          getEvents(
-            term,
-            currentPage,
-            perPage,
-            order,
-            orderDir,
-            filters,
-            extraColumns
-          )
-        );
-      })
-      .catch(() => {
-        console.log("ERROR");
-      });
-  };
+  return putRequest(
+    null,
+    createAction(UPDATED_REMOTE_EVENTS)({}),
+    `${window.API_BASE_URL}/api/v1/summits/${currentSummit.id}/events/?access_token=${accessToken}`,
+    {
+      events: normalizedEvents
+    },
+    authErrorHandler
+  )({})(dispatch)
+    .then(() => {
+      dispatch(stopLoading());
+      dispatch(
+        showSuccessMessage(
+          T.translate("bulk_actions_page.messages.update_success"),
+          () => history.push(`/app/summits/${currentSummit.id}/events/`)
+        )
+      );
+      const {
+        currentEventListState: {
+          term,
+          currentPage,
+          perPage,
+          order,
+          orderDir,
+          filters,
+          extraColumns
+        }
+      } = getState();
+      dispatch(
+        getEvents(
+          term,
+          currentPage,
+          perPage,
+          order,
+          orderDir,
+          filters,
+          extraColumns
+        )
+      );
+    })
+    .catch(() => {
+      console.log("ERROR");
+    });
+};
 
 export const getActionTypes =
   (selectionPlanId) => async (dispatch, getState) => {
